+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 0.9.3
+- Small bugfixes
 
 ## 0.9.2
 - Show/Edit passwords (+ delete passwords -> move to trash)