+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 - Add settings screen:
     - Set start view: All Passwords, Folders or Favorites
     - Enable/Disable biometric auth
-    
+- Small bugfixes
+
 ## 0.9.2
 - Show/Edit passwords (+ delete passwords -> move to trash)
 - Show folders (+ create new Folders & rename existing folders)